refactor(0.0.0): Separate BotConfig classes
